make sure the occlusion maps benefit from the same kind of optimization as the shadowmap could have. Mainly each shadowmap composing an occlusion map store its minimum value. Later when occlusion() is called it looks first if the min. value was saved for a certain shadowmap and discard quickly the sample which could have been in shadow.
make sure the display compute the min value as double value (to save in the tiff file as SMinSample).

make sure if you use either MakeShadow, MakeOcclusion or output directly as "shadow" the minimum value for a shadow is saved correctly into the shadow tiff file or occlusion tiff file.


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@72 30f9f022-c70d-0410-a6f5-8c645ca7840b