This is to speed up enormously occlusion() 256x256x256. 	<aqsis/aqsis/content/ribs/features/occlusion> example
I change a bit CqTextureMap class to contain two type of list of segments.
One list as before (m_apFlat) which is straight list of CqTextureMapBuffer.
Second list is an array of 256 lists which organized the tiles according to the directory id. (index).
In case of occlusion each lists will have only the segment for a certain shadowmap. cubic environment will hold roughly 6 lists (6 faces).
In case of mipmaps (regular texture) about the log2() of the min(width,height )

I simplify a lot the ::GetBuffer() to kept.


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@77 30f9f022-c70d-0410-a6f5-8c645ca7840b