Fix the lighting problem with lightsources defined outside the WorldBegin/End block. The problem is that at definition time, the shader's idea of 'camera' space is identity, while at 'preparation' time, it is different, it is the real camera space either that defined as the camera in the RIB (i.e. at the time of RiWorldBegin) or that defined during a custom render pass, like autoshadows.
As this can only possibly affect lightsources, because it is illegal to define other shader types
outside the world block, we now check if the lightsource is being defined outside the world block
at definition time, and if so, just go ahead and initialise the parameters now, as they are 
invariant under any camera space changes applied during and auto* pass. In addition we store
with a shader whether it was defined outside the world block, and if so, ignore the request
to 'prepare' the shader, as it is already properly prepared.


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@172 30f9f022-c70d-0410-a6f5-8c645ca7840b