Apply patches from Chris Foster to implement matrix multiplication and division services for the shading language, and to reformat the matrix output from the SL printf function to be more usable.
------------------------

Here's another patch.  This one provides matrix multiplication and
division services for the shading language.

I've tested it by:
1) running the regression tests against an untouched copy of the latest
svn - they both produce the same output
2) writing a shader to spit out the results of M1*M2 and M1/M2 via
printf, and compare to an external program (cgkit & matlab)

The file is matrix_shadeops.patch

I'm attaching another patch as well (printf_matrix_format.patch).  This
one changes the formatting which the printf function gives to matrices
in the shader language.  In particular, it prints matrices on a single
line which is IMHO much easier to work with, particulary since the
output can be filtered by unix utilities such as sort and uniq.
3delight also uses a single line format.



git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@266 30f9f022-c70d-0410-a6f5-8c645ca7840b