Fix bug #1471457 :  Can't use atmospheric and displacement shaders together.
There was a test in the CqSurface::Uses() function that was supposed to check for the absence of any shader assignment, and in that case return 0. However, the test was faulty, and in fact checked for the _presence_ of all three and returned 0.

Added a test case for the above bug to the RTS.


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@276 30f9f022-c70d-0410-a6f5-8c645ca7840b