I speed up the creation of any parameters within aqsis. Parameters are a key elements. Each attributes, options and parameters attached to any primitives are found easier and faster and using less memory.
I choose the following when we are looking for a parameter we are really interessed if two parameters's name match. Eg. you loop into all the variable to match to a certain parameters names like "st". I used to compared only adress of string and kept track of each strings (for a parameter) into a local table of strings with paramaters.cpp. 

A lot of code was changed because instead of saving a m_strName as CqString I used a simple char * where ultimely pointed as mentionned earlier to a table of strings. The table of strings (class) are defined with sstring.cpp, sstring.h. We could use as many table of strings as required but remember the table are meant to be found an equal string it is no way a relational database. You do a CqStringTable::Get() to get a new string, ::Dump to print to the console, ::Wipe() to delete the strings entries.



git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@320 30f9f022-c70d-0410-a6f5-8c645ca7840b