This update includes minor GUI updates, a BtoR spacehandler script to pass object selections back to BtoR, and the BtoRMain.py script, the central part of BtoR.
This contains the working material editor, with save/restore & preview ability.

The material editor preview probably has a few scale issues, so that needs to be reworked. I'm probably going to export something from Rhino that's to a more appropriate scale for blender.

Note that exception handling for problematic shaders remains to be added before this is considered stable.



git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@339 30f9f022-c70d-0410-a6f5-8c645ca7840b