Enable the use of the MPDump facility in the release build. On by default, can be disabled at build time by passing 'enable_mpdump=false' to SCons.
When enabled, passing '-mpdump' to aqsis will cause a file 'mpdump.mp' to be generated containing a dump of micropolygons and other data generated by Aqsis during rendering. This file can be processed using the mpanalyse.py tool.

The '-mpdump' command line option basically sets the option 

    Option "mpdump" "integer enabled" [1]

so setting this Option in the RIB will have the same effect, 1 is enabled, 0 is disabled.

git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@530 30f9f022-c70d-0410-a6f5-8c645ca7840b