Made one step closer to eliminating the splotchiness problem in the deep data: now it only occurs where semi-transparent geometry borders opaque geometry.
Added worldToScreen and worldToCamera transformation matrices to SqDtexFileHeader, and added the necessary lines to forward those matrices from the renderer, down the pipeline and into the dtex file.

I output my first *.dsm file using a resolution of 400x300 and a pixel filter size of 2 and 2 pixel samples, and the (uncompressed) dtex file is over 20MB. Pretty big!



git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/branches/soc2007-dsm/aqsis@1468 30f9f022-c70d-0410-a6f5-8c645ca7840b