Make sure a huge memory leak in CqPool is plugged for all g++ compiler versions.
Originally I observed the issue (see comments in code) with g++ version 3.4,
but not with 4.1; I imagined that it was possibly a very subtle compiler bug
which had been fixed in version 4.  Now I have observed the leak with g++-4.2
as well, so it seems safest to enable the fix for all versions of gcc.


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2248 30f9f022-c70d-0410-a6f5-8c645ca7840b