Try to fix an odd build failure on gentoo, as reported at http://bugs.gentoo.org/show_bug.cgi?id=230171
We currently link to the dl library for dlopen(), dlclose() etc for each
executable separately, but there doesn't seem to be a good reason for doing so.
Instead I've chosen to link aqsistypes with the dl library; that dependency
should then carry through to any executables which need it.  This was done by
introducing a variable - AQSISTYPES_SYSTEM_LINKLIBS - and adding dl to it in
the platform-specific cmake files.  Also I cleaned up a similar WIN32-specific
library using the same variable.



git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2255 30f9f022-c70d-0410-a6f5-8c645ca7840b