Bugfix: Fix a serious memory leak involving string arguments to DSO shadops.
Storage space for string arguments to a DSO shadops was allocated but never
freed.  This caused a serious leak when using any DSO shadeop with string
arguments.

I've also made sure that we allocate memory for use in STRING_DESC structures
using malloc() since the DSO interface is C, and (presumably) DSOs may want to
free() the memory and malloc a larger block.


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2341 30f9f022-c70d-0410-a6f5-8c645ca7840b