Bugfix: Ensure that AQSIS_FLTK_INCLUDE_DIR is only used when AQSIS_USE_FLTK is true.
This prevents a cmake makefile generation failure in the case that
AQSIS_USE_FLTK is true, but fltk is not found on the system.

Also reformat two status messages for emphasis and accuracy - display *will* be
built when FLTK isn't found, but the the legacy framebuffer part of it won't.


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2359 30f9f022-c70d-0410-a6f5-8c645ca7840b