Implement importance sampling for the new occlusion mapping code.
As a result, samples are spent where they are needed most, rather than in
sampling shadow maps which lie at a glancing angle to the surface.  This should
result in significantly reduced noise for a given number of samples.

The algorithm used also results in correct number of occlusion samples being
taken, corresponding to the "samples" parameter to the occlusion() shadeop.
For this to be accurate, the shadow maps must be well-distributed over the
sphere.



git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2379 30f9f022-c70d-0410-a6f5-8c645ca7840b