Implement binary RIB decoding in the new rib lexer.
Remaining TODO's indicate some areas which still need investigation.  In
particular, the old binary decoder seemed to decode fixed point integers in a
way contrary to the spec - need to investigate whether this was intentional or
not.  I've also decided to return the ENDOFFILE token when "\377" or "#\377" is
read, and need to check that this is correct.

The changes are partially tested with test cases, though I still need to add
tests for encoded requests, strings, floats and float arrays.

Additional stuff:
* Add support types TqUint64 and TqInt64 to aqsis_types.

* Minor changes to CqRibToken to allow access to the held string.  This is an
  attempt to avoid copying the string as much as possible.



git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/branches/new_ribparser/aqsis@2411 30f9f022-c70d-0410-a6f5-8c645ca7840b