Fixed a bug in the binary RIB fixed-point decoding.
The code incorrectly interpreted the RISpec to mean that the number of
fractional bytes was added to the number of magnitude bytes.  Instead, it is
the position of the radix point which is specified, along with the total number
of bytes.

Thanks to Paul for testing.  The new expected behaviour should probably be
tested against ribcat as well.


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/branches/new_ribparser/aqsis@2412 30f9f022-c70d-0410-a6f5-8c645ca7840b