Add a utililty for enum constant <---> string conversions.
The problem of converting a string to an enum value and vice versa comes up
time and time again.  Each time we're faced with implementing an ad-hoc
solution which may not be efficient and typically involves an array of strings
in a different file.  Typically such solutions are fragile and hard to
maintain.

This commit contains a reasonably general solution to the problem, and should
be considered for all future enum to string conversions.  Some problems which
remain include
  * CqEnum can't handle enums with non-default values.  This can be remedied in
    the future if necessary.
  * The macros AQSIS_ENUM_INFO_BEGIN / AQSIS_ENUM_INFO_END won't work outside
    the Aqsis namespace.


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2422 30f9f022-c70d-0410-a6f5-8c645ca7840b