Improve piqsl responsiveness by retrying piqsl socket connection at 0.1s intervals.
This replaces the previous fixed behaviour of waiting for 2 seconds to allow
piqsl to start up, and results in a much more responsive feel for the
application.  The timeout is set for 10s, so should also work better on very
slow machines.

The unix sleep() function only works in second intervals, so had to be replaced
with the POSIX nanosleep() function from time.h


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2473 30f9f022-c70d-0410-a6f5-8c645ca7840b