Remove the AQSIS_USE_PLUGINS option from cmake to avoid future confusion and broken builds.  Currently plugins are required to build a usable version of aqsis, since the rendered data can only come out of the renderer via the plugin Dspy interface.
Also changed the PLUGINS macro, used inside plugins.cpp for turning on
compilation of plugins.  The macro is now changed to AQSIS_NO_PLUGINS, which
only needs to be defined in some strage case where you might not want plugins
compiled.  This means that all mention of plugins could be removed from the
build system.


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2490 30f9f022-c70d-0410-a6f5-8c645ca7840b