Bugfix #2167498: "uniform user parameters not working for SDS"
This was due to an error in CqSurfaceSubdivisionPatch::StoreDiceAPVar(), in
which the varying index was used for all primvar classes except facevarying and
facevertex.  Adjusting the function to use CqSurfaceSubdivisionPatch::m_FaceIndex
for class_uniform primvars fixes the problem.  A similar problem for
class_constant was also fixed with this patch.


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2539 30f9f022-c70d-0410-a6f5-8c645ca7840b