Fix a TODO which was causing hairgen to crash on windows (thanks Paul :)
Uniform and constant primvars attached to the parent mesh weren't being
copied onto the child hairs properly.  What's more, these could cause hairgen
to read of the end of an array if they were inserted in the middle of the
parameter list.

Also moved the position of AQSIS_EXPORT to allow proper compilation on windows
(bah, I got this wrong again; thanks to Paul for reporting).


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2544 30f9f022-c70d-0410-a6f5-8c645ca7840b