Merge r2497,r2676 - GeometricApproximation "focusfactor" optimisations.
The changes in r2676 interacted with the changes merged from the multithreading
branch, which made some conflicts when merging.  In particular, the SqGridInfo
struct is now handled in a completely different way.

A SqGridInfo2 struct was created to mirror the new usage of SqGridInfo in the
trunk and enable the patch without major changes.


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/branches/aqsis-1.4/aqsis@2704 30f9f022-c70d-0410-a6f5-8c645ca7840b