Bugfix: Fix black spots due to environment maps in MOSAIC example scene.
Black spots appear in the image from the MOSAIC mapping tutorial due to the
environment filtering code producing NaN's which then propagated via the
filtering operation.

This was due to a complex sequence of events:

1. The refraction() built-in function returns (0,0,0) for total-internal
   reflection.  Unfortunately this is bad from a filtering point of view,
   since some points on a grid have a valid reflection vector, and others
   have (0,0,0).  MOSAICsurface.sl implicitly calls refraction() from the
   fresnel() built-in.
2. For shading points which have non-zero transmission (T) vector, but have two
   adjacent shading points where T_ij = (0,0,0), the derivatives come out as -T
   in both directions.
3. The reconstructed quadrilateral inside SampleMap is then T1 = T, T2 = 0,
   T3 = 0, T4 = -T.
4. The first sample direction for a cubeface environment is always 
   the centre point 1/4*(T1 + T2 + T3 + T4) = (0,0,0).
5. Sampling the environment map at (0,0,0) doesn't make sense and results in a
   NaN due to division by zero.
6. NaN's pollute any further float arithmetic operations, including the
   filtering which results in black spots when converted to integer color
   channels.

The workaround makes the code more robust by extracting the "true" filtering
quadrilateral directly from the grid rather than approximating it with a pair
of derivatives.  With the fix it's extremely unlikely that the unfortunate
add-to-zero coincidence in the previous code will occur, unless the user puts
in a full grid of zero environment vectors.

This bug doesn't appear in the 1.5 branch since the environment map code has
been completely rewritten.


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/branches/aqsis-1.4/aqsis@2706 30f9f022-c70d-0410-a6f5-8c645ca7840b