Bugfix #2531309: artifacts with book export in piqsl
This was due to libtiff trashing the in-memory buffer during image compression.
The first time the image is saved the buffer held by piqsl was messed up by
libtiff.  Saving a second time caused the rubbish data to be saved to the file.

The simple fix is to copy the data to a temporary scanline-sized buffer before
calling TIFFWriteScanline(), and pass the copy rather than the original data.


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2725 30f9f022-c70d-0410-a6f5-8c645ca7840b