Bugfix #1829339: Artefacts in SDS render (large micropolygons)
The Diceable() check for subdivision surfaces had difficulty calculating the
correct dice size for some mesh geometries.  For subdivision meshes which have
a large polygon next to a very small one, the size of the small polygon on the
limit surface can be larger by an arbitrary amount.  However, the Diceable()
function used the positions of the non-limit mesh vertices to estimate the
dicing rate for a subdivision patch.  This sometimes led to huge underestimates
of the appropraite dice size with very large micropolygons as a result.

The solution taken here is to make sure that vertices are pushed to the limit
surface before being used in the dice rate estimation.  This is done using the
newly implemented limitPoint() function which makes use of limit masks.  The
current code doesn't yet deal correctly with edge hardness, but gives far
better results than the old code in other cases.

Other minor changes:
  * Made various functions in CqLath and CqSubdivisionPatch const-correct.
  * Created CqLath::isCornerVertex()


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2731 30f9f022-c70d-0410-a6f5-8c645ca7840b