Bugfix #2013783: Large cracks in subdivision surfaces
This bug was due to the difference between SDS bspline dicing vs dicing via
recursive subdivision.  In particular, the subdivided grid doesn't quite lie on
the limit surface whereas the bspline grid does - this can lead to large
cracks in some cases.   The solution is to apply a limit mask to the subdivided
points which forces them onto the limit surface, closely matching any adjacent
bspline grids.

Pushing points to the limit surface has a significant performance hit when
using facevertex variables (up to 50% or so in subdivision-bound scenes).  This
needs to be mitigated by converting quadratic patch neighbourhoods to bsplines
in more cases than we currently do.

This patch also contains other modifications and fixes to the subdivision code:

* Fix for the recently introduced limitPoint() function.  This function
  requires that all neighbouring parent patches are subdivided in order to
  apply the limit mask correctly, but this wasn't the case before.

* Improved the dice resolution estimation.  Previously the renderer would often
  generate micropolygons which were larger than the shading rate.  This has
  been corrected so that only sufficiently small micropolygons are generated.
  As expected this makes subdivision dicing slower, but the user is always free
  to adjust the quality/performance tradeoff using the shading rate...




git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2733 30f9f022-c70d-0410-a6f5-8c645ca7840b