Reimplement the SqSampleData storage to use boost::intrusive_ptr.
The CqBucketProcessor creates them and stores a reference to them
in a vector of intrusive_ptr's. CqImagePixel then stores it's own
references. 

This will allow the caching mechanism to just transfer ownership
from one bucket to another. 

Also refactored the image pixel code a bit, cleaning up the code
to follow the standard, and simplified the SqImageSample struct
to make it a simple lightweight structure.



git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2736 30f9f022-c70d-0410-a6f5-8c645ca7840b