Create IqTextureCache, an interface for CqTextureCache.
The aim of this interface is to allow proper decoupling between
libaqsis_shadervm and libaqsis_tex to avoid requiring the shadervm to link to
the texture library.  The corresponding build change was made in r2746, but
resulted in a broken build with undefined symbols on OSX (and undoubtably
windows too).

Details:

* Introduce IqTextureCache, modelled on the CqTextureCache interface.
* Make sure the texture shadeops only interact with the abstract interface
  IqTextureCache by modifying IqRenderer to return the abstract class.
* Make the simple class CqTextureSampleOptions header-only
* Remove some old texture stuff from the texture shadeops file.



git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2748 30f9f022-c70d-0410-a6f5-8c645ca7840b