Fix a bug in the new RIB variable parameter list handling.
The string arguments were held in a vector of vectors which was subject to
reallocation as the set of stored string arrays increased in size.  The fix
avoids the resulting dangling pointers by holding the vectors via shared_ptr so
that they don't get reallocated.

Also added a unit test testing the behaviour.


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2751 30f9f022-c70d-0410-a6f5-8c645ca7840b