Fix a bug in RI parameter list handling which meant that the token count() function sometimes returned zero.
This was because count() relies on the size of m_tokens which is filled only
when tokens() is called.  The bug wasn't apparent on all architectures because
the order of evaluation of terms in a function call (where tokens() and count()
are typically used) is undefined.

Also updated some test cases to hopefully cover the bug in future.


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2778 30f9f022-c70d-0410-a6f5-8c645ca7840b