Integrate the new RIB parser with RiProcRunProgram support and introduce popen-like functionality to aqsistypes to localise the OS-dependent code in one place.
The windows implementation of CqPopenDevice is a stub in this commit, so
RiProcRunProgram functionality will be missing on windows for a short while
until this is fixed.  There's two additional issues which need to be cleaned
up: 1) Procedurals aren't correctly found in the current directory even if it's
part of the search path, and 2) The program repository is a global whereas it
should be a member of the CqRenderer class.

Details:

* Introduce a new class CqPopenDevice which works with boost::iostreams::stream
  to provide bidirectional popen-like functionality in a std::iostream framework.
* Rearrange aqsistypes/CMakeLists.txt to explicitly state all source files for
  finer-grained control over the way test files are dealt with.
* Add CqRunProgramRepository class to manage all current RunProgram instances
  in a system-independent way using streams constructed from CqPopenDevice.
* Refactor RiProcRunProgram to use CqRunProgramRepository and new parser.
* Add findFileInRiSearchPath to replace use of CqRiFile as a file search
  mechanism.


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2781 30f9f022-c70d-0410-a6f5-8c645ca7840b