Bugfix - vase.rib entered an infinite loop on some architectures.
This bug was due to incorrect code in CqSurfacePatchBicubic::Diceable()
which intended to disallow dicing when a patch was too curved.  Due to
normalising both vec and vecL in the removed loops, the "distance to the line"
was actually measuring sin of the angle between vecL and vec instead.  This
meant that the code almost never resulted in diceable rejection as intended.
A bug in the inner loops which caused iteration off the array end - coupled
with floating point inaccuracy - were the cause of the infinite splitting loop
for vase.rib.

This patch simply removes the offending code, since the diceable check which
remains appears to work well enough.  In any case, a Bezier curve always lies
within the convex hull of its control points.  This suggests that the
calculated lengths uLen and vLen are actually fine for estimating the dicing
rate.


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2801 30f9f022-c70d-0410-a6f5-8c645ca7840b