Optimizations for the sample hit storage.
Sample hits were stored in SqSampleData using a std::deque, but std::vector is
significantly smaller and faster.  The depthcomplexity tests in the
performance test suite indicate that this is a useful optimization, with a
speedup of a factor of two in some cases.

In addition, the CqBucketProcessor::StoreSample() function was performing
unnecessary work: It sorted and inserted samples on the fly which implies an
O(N^2) algorithm.  Following that, it sorted them in CqImagePixel::Combine(),
which was completely unnecessary.  Instead, we now leave samples unsorted
during sampling, and only sort them during Combine(), leading to an O(N
log(N)) algorithm.  This optimisation was actually included in revision 297,
but somehow lost again, probably when we reverted to the old scanline
rasterizer code during the multithreading merge.


Details:
--------

* Store sample hits (SqSampleData::data) in a std::vector rather than a
  std::deque for efficiency.

* Refactor CqBucketProcessor::StoreSample() to remove the linear search and
  push samples directly onto the end of the storage to be sorted later.
  Remove a unreliable check against hitting "micropolygon grid lines".  As
  noted in r297, this isn't reliable (due to floating point error) and there's
  better ways to do this anyway.

* Rename SqImageSample::opaqueSample to occludingHit for accuracy: this
  variable may in fact only hold a subset of the opaque micropoly hits.  For
  instance, CSG nodes need to be stored in the hit list even when they're
  opaque.  Related renaming of CqImagePixel::OpaqueValues().

* Remove SqMpgSampleInfo.isOpaque since it's no longer necessary.

* Remove SqImageSample::Flag_Occludes since it's no longer used.

* Change SqImageSample::flags to unsigned for cleanliness since it's a
  bitfield.

* Remove some #includes in imagepixel.{h,cpp}

* Remove m_ prefix from SqGridInfo members to follow the coding convention.

* Add useSmoothShading member to SqGridInfo to avoid many lookups of
  the ShadingInterpolation attribute.



git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2825 30f9f022-c70d-0410-a6f5-8c645ca7840b