Commit current implementation of the Beaker:BOC project.   * The caching is now split into 8 segments,     top-left, top, top-right, left, right, bottom-left, bottom & bottom-right.   * The corner segments may be referenced from multiple buckets.   * The CqBucket class stores the segments in a fixed length array     of boost::shared_ptrs.   * To avoid the situation where CqImagePixels exist in multiple     places in the bucket processor pixel array, any pixels that were     introduced from a cache segment are replaced when the bucket is     finished from the pixel pool, which shouldn't be too expensive.
Currently the code is a little slower than without caching, but it is
reducing the sample count significantly. The implementation needs review to find
the current bottlenecks in the code.

Note: the function CqBucketProcessor::setupCacheInformation is badly named, as
it now sets up general static information used by the processor(s) throughtout
the rendering, but I couldn't think of a better name, so it stays for now.



git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2831 30f9f022-c70d-0410-a6f5-8c645ca7840b