Optimisations to occlusion tree - make every leaf node have exactly one sample.
Previously, the occlusion tree was set up to span the entire sample region of
the current bucket.  However, since it's a balanced binary tree, if the number
of samples in the x and y directions aren't power-of-two then the leaf nodes of
the tree don't contain exactly one sample.  This leads to several complications
and inefficiencies, including reduced cache locality when updating the leaf
nodes with a new sample depth.

This patch changes the region that the tree covers so that each leaf node of
the tree corresponds to exactly one or zero samples.  To deal with the
power-of-two missmatch, we let parts of the tree "hang off" the edge of the
bucket unused.

Details:
* Many simplifications to the occlusion tree code due to the new
  single-sample-per-leaf design.
* Introduce an iterator for the samples held in a CqBucketProcessor to
  decouple the occlusion tree implementation from the details of the bucket
  processor sample storage scheme.
* Introduce some tests for the tricky position -> leaf index calculation in the
  internals of the occlusion tree.


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2838 30f9f022-c70d-0410-a6f5-8c645ca7840b