Remove the subCellIndex member from SqSampleData
The index for the multijittered subcell was necessary for the old method of
filtering - it tried to cache the filter values more accurately for the special
case of the multijittered sampling pattern.

A generic sample point interface was added in r2835, so the sub cell index is
no longer relevant.  Filter values are now simply cached at the centre of the
subpixels, which is slightly more efficient anyway and seems to give perfectly
reasonable results.

Also moved a check for a valid filter function from InitialiseFilterValues to
RiPixelFilter.


git-svn-id: https://aqsis.svn.sourceforge.net/svnroot/aqsis/trunk/aqsis@2839 30f9f022-c70d-0410-a6f5-8c645ca7840b