Make sure the slxargs library correctly reports shader variable storage types.
* Replace IqShaderData::fParameter() with IqShaderData::Storage(), and create a
new enum, IqShaderData::EqStorage.  This allows us to query the "storage
type" of a shader variable - that is, whether it's a local temporary on the
shader, a parameter to the shader, an output parameter etc.

* Various changes to the shaderVM etc as necessary to account for the change to
Storage().

* Add new enum values SLX_STOR_OUTPUTPARAMETER and SLO_STOR_OUTPUTPARAMETER to
the slx and slo interfaces respectively, and hook them up to the Storage()
method.

* Hook aqsltell up to Slx_StortoStr() so that the storage type is correctly
reported.

* Bugfix: Fix a crash in GetCurrentShaderInfo which arose whenever the DSO path
wasn't correctly set.